+++ app/(main)/page.tsx
@@ -1,5 +1,6 @@
-import Feed from "@/components/pages/Feed";
+import Feed from "@/components/pages/main/Feed";
 import ChatCardList from "@/components/pages/main/ChatCardList";
+
 export default function Home() {
   return (
     <>
